+++ website/src/lib/theme/theme-typography.ts
@@ -7,17 +7,21 @@ import type { ThemeTypography } from '@uui';
  *   <Typography element="div" variant="body3">Custom Typography</Typography>
  * </Theme>
  */
+
 const customTypography: ThemeTypography = {
   body3: {
     fontWeight: 'bold',
     fontSize: 'var(--h6-font-size)',
     lineHeight: '143%',
-    fontStyle: 'normal',
     fontFamily: 'cursive',
+    fontStyle: 'normal',
     letterSpacing: '.17px',
     textTransform: 'uppercase',
     textDecoration: 'underline wavy var(--warning)',
-    color: 'var(--primary)'
+    color: 'var(--primary)',
+    WebkitTextDecorationLine: 'underline',
+    WebkitTextDecorationStyle: 'wavy',
+    WebkitTextDecorationColor: 'var(--warning)'
   }
 };