+++ website/src/routes/scoped-styles/page-styles.ts
@@ -1,0 +1,12 @@
+import type { ScopedStyles } from '@uui';
+
+export const scopedExample: ScopedStyles = {
+  $self: {
+    gap: '32px',
+    width: '100%',
+    display: 'flex',
+    flexDirection: 'column',
+    alignItems: 'center',
+    justifyContent: 'center'
+  }
+};